Simplify: defer pyyaml, dedupe trim_timestamp + test fixtures
- Defer ``import yaml`` to inside the ``-o yaml`` branch in render_output. yaml import is a measurable chunk of CLI startup and only the rare yaml path needs it.
- Promote log_format._trim_timestamp to public ``trim_timestamp`` and drop the duplicate copy in textual/deployment_monitor.py.
- Hoist parse_log_body/render_plain import in commands/deployment.py to module top, the prior local import was paid per log event.
- Move shared test fixtures (make_deployment, fake_profile, patched_auth, patch_project_client) to tests/conftest.py and use them from both deployments_get_output and deployments_logs_status tests.
- Strip internal "Slice A/B" / thoughts/ references from public-repo files.
- Rename PascalCase Auth_dataclass_factory to _make_auth_profile.
