+++ packages/llamactl/tests/test_deployments_get_output.py
@@ -1,11 +1,7 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2026 LlamaIndex Inc.
 """Tests for ``deployments get`` output modes, ``--project`` override, and the
-``deployments list`` hidden alias.
-
-Phase 1 of the llamactl Slice A redesign — see
-``thoughts/shared/plans/2026-04-26-llamactl-slice-a-output-modes-logs-status.md``.
-"""
+``deployments list`` hidden alias."""
 
 from __future__ import annotations
 
@@ -15,9 +11,9 @@ from types import SimpleNamespace
 from typing import Any
 from unittest.mock import MagicMock, patch
 
-import pytest
 import yaml
 from click.testing import CliRunner
+from conftest import make_deployment, patch_project_client
 from llama_agents.cli.app import app
 from llama_agents.core.schema.deployments import (
     DeploymentHistoryResponse,
@@ -26,62 +22,15 @@ from llama_agents.core.schema.deployments import (
 )
 
 
-def _make_deployment(deployment_id: str = "my-app", **overrides: Any) -> DeploymentResponse:
-    base: dict[str, Any] = {
-        "id": deployment_id,
-        "display_name": deployment_id,
-        "repo_url": "https://github.com/example/repo",
-        "deployment_file_path": "llama_deploy.yaml",
-        "git_ref": "main",
-        "git_sha": "abc1234567890",
-        "project_id": "proj_default",
-        "secret_names": ["FOO", "BAR"],
-        "apiserver_url": None,
-        "status": "Running",
-    }
-    base.update(overrides)
-    return DeploymentResponse.model_validate(base)
-
-
-@pytest.fixture
-def fake_profile() -> SimpleNamespace:
-    return SimpleNamespace(
-        api_url="http://test:8011",
-        project_id="proj_default",
-        api_key="key",
-        device_oidc=None,
-        name="prof",
-    )
-
-
-@pytest.fixture
-def patched_auth(fake_profile: SimpleNamespace):
-    """Patch the env service so commands use a fake authenticated profile."""
-    with patch("llama_agents.cli.config.env_service.service") as mock_service:
-        mock_auth_svc = MagicMock()
-        mock_auth_svc.get_current_profile.return_value = fake_profile
-        mock_auth_svc.list_profiles.return_value = [fake_profile]
-        mock_auth_svc.env = SimpleNamespace(requires_auth=True)
-        mock_auth_svc.auth_middleware.return_value = None
-        mock_service.current_auth_service.return_value = mock_auth_svc
-        yield mock_service
-
-
-def _patch_project_client(client_mock: MagicMock):
-    """Patch ProjectClient construction inside ``cli.client.get_project_client``."""
-    return patch(
-        "llama_agents.core.client.manage_client.ProjectClient",
-        return_value=client_mock,
-    )
-
-
 def _make_client_mock(deployments: list[DeploymentResponse]) -> MagicMock:
     """A mock ProjectClient stand-in with the methods the commands hit."""
 
     async def _list() -> list[DeploymentResponse]:
         return list(deployments)
 
-    async def _get(deployment_id: str, include_events: bool = False) -> DeploymentResponse:
+    async def _get(
+        deployment_id: str, include_events: bool = False
+    ) -> DeploymentResponse:
         for d in deployments:
             if d.id == deployment_id:
                 return d
@@ -97,9 +46,9 @@ def _make_client_mock(deployments: list[DeploymentResponse]) -> MagicMock:
 
 def test_deployments_get_text_no_args_lists(patched_auth: Any) -> None:
     runner = CliRunner()
-    deployments = [_make_deployment("app-a"), _make_deployment("app-b")]
+    deployments = [make_deployment("app-a"), make_deployment("app-b")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(app, ["deployments", "get", "--no-interactive"])
     assert result.exit_code == 0, result.output
     assert "app-a" in result.output
@@ -108,9 +57,9 @@ def test_deployments_get_text_no_args_lists(patched_auth: Any) -> None:
 
 def test_deployments_get_json_array(patched_auth: Any) -> None:
     runner = CliRunner()
-    deployments = [_make_deployment("app-a"), _make_deployment("app-b")]
+    deployments = [make_deployment("app-a"), make_deployment("app-b")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "get", "--no-interactive", "-o", "json"]
         )
@@ -124,9 +73,9 @@ def test_deployments_get_json_array(patched_auth: Any) -> None:
 
 def test_deployments_get_yaml_list(patched_auth: Any) -> None:
     runner = CliRunner()
-    deployments = [_make_deployment("only-one")]
+    deployments = [make_deployment("only-one")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "get", "--no-interactive", "-o", "yaml"]
         )
@@ -139,10 +88,10 @@ def test_deployments_get_yaml_list(patched_auth: Any) -> None:
 def test_deployments_get_single_text_no_tui(patched_auth: Any) -> None:
     """``deployments get <name>`` should never launch the Textual monitor."""
     runner = CliRunner()
-    deployments = [_make_deployment("my-app")]
+    deployments = [make_deployment("my-app")]
     client_mock = _make_client_mock(deployments)
     with (
-        _patch_project_client(client_mock),
+        patch_project_client(client_mock),
         patch(
             "llama_agents.cli.textual.deployment_monitor.monitor_deployment_screen"
         ) as mock_monitor,
@@ -156,9 +105,9 @@ def test_deployments_get_single_text_no_tui(patched_auth: Any) -> None:
 
 def test_deployments_get_single_json(patched_auth: Any) -> None:
     runner = CliRunner()
-    deployments = [_make_deployment("my-app")]
+    deployments = [make_deployment("my-app")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "get", "my-app", "--no-interactive", "-o", "json"]
         )
@@ -171,9 +120,9 @@ def test_deployments_get_single_json(patched_auth: Any) -> None:
 
 def test_deployments_get_single_yaml(patched_auth: Any) -> None:
     runner = CliRunner()
-    deployments = [_make_deployment("my-app")]
+    deployments = [make_deployment("my-app")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "get", "my-app", "--no-interactive", "-o", "yaml"]
         )
@@ -186,7 +135,7 @@ def test_deployments_get_single_yaml(patched_auth: Any) -> None:
 def test_deployments_get_empty_json_is_array(patched_auth: Any) -> None:
     runner = CliRunner()
     client_mock = _make_client_mock([])
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "get", "--no-interactive", "-o", "json"]
         )
@@ -197,9 +146,9 @@ def test_deployments_get_empty_json_is_array(patched_auth: Any) -> None:
 def test_deployments_list_hidden_alias_works(patched_auth: Any) -> None:
     """The hidden ``deployments list`` alias should still produce JSON output."""
     runner = CliRunner()
-    deployments = [_make_deployment("app-a")]
+    deployments = [make_deployment("app-a")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app, ["deployments", "list", "--no-interactive", "-o", "json"]
         )
@@ -222,9 +171,9 @@ def test_deployments_get_project_override_threads_to_client(
 ) -> None:
     """``--project foo`` should construct a ProjectClient with project_id='foo'."""
     runner = CliRunner()
-    deployments = [_make_deployment("app-a", project_id="proj_other")]
+    deployments = [make_deployment("app-a", project_id="proj_other")]
     client_mock = _make_client_mock(deployments)
-    with _patch_project_client(client_mock) as ctor:
+    with patch_project_client(client_mock) as ctor:
         result = runner.invoke(
             app,
             [
@@ -258,7 +207,7 @@ def test_deployments_history_json_output(patched_auth: Any) -> None:
     async def _hist(deployment_id: str) -> DeploymentHistoryResponse:
         return DeploymentHistoryResponse(deployment_id=deployment_id, history=items)
 
-    client_mock = _make_client_mock([_make_deployment("my-app")])
+    client_mock = _make_client_mock([make_deployment("my-app")])
     client_mock.get_deployment_history.side_effect = _hist
     client_mock.aclose = MagicMock()
 
@@ -266,7 +215,7 @@ def test_deployments_history_json_output(patched_auth: Any) -> None:
         return None
 
     client_mock.aclose.side_effect = _aclose
-    with _patch_project_client(client_mock):
+    with patch_project_client(client_mock):
         result = runner.invoke(
             app,
             [
@@ -286,25 +235,7 @@ def test_deployments_history_json_output(patched_auth: Any) -> None:
     assert data[1]["git_sha"] == "aaaaaaa1111"
 
 
-def test_auth_list_json_omits_secrets(patched_auth: Any) -> None:
-    runner = CliRunner()
-    with patch("llama_agents.cli.config.env_service.service") as mock_service:
-        mock_auth_svc = MagicMock()
-        profile = Auth_dataclass_factory()
-        mock_auth_svc.list_profiles.return_value = [profile]
-        mock_auth_svc.get_current_profile.return_value = profile
-        mock_auth_svc.env = SimpleNamespace(requires_auth=True)
-        mock_service.current_auth_service.return_value = mock_auth_svc
-        result = runner.invoke(app, ["auth", "list", "-o", "json"])
-    assert result.exit_code == 0, result.output
-    data = json.loads(result.output)
-    assert isinstance(data, list)
-    assert data[0]["name"] == "prof"
-    # Secrets must not leak through.
-    assert "api_key" not in data[0]
-
-
-def Auth_dataclass_factory() -> Any:
+def _make_auth_profile() -> Any:
     """Build a real ``Auth`` dataclass; helpers like SimpleNamespace fail
     equality checks used by the active-profile indicator."""
 
@@ -319,6 +250,24 @@ def Auth_dataclass_factory() -> Any:
         api_key_id=None,
         device_oidc=None,
     )
+
+
+def test_auth_list_json_omits_secrets(patched_auth: Any) -> None:
+    runner = CliRunner()
+    with patch("llama_agents.cli.config.env_service.service") as mock_service:
+        mock_auth_svc = MagicMock()
+        profile = _make_auth_profile()
+        mock_auth_svc.list_profiles.return_value = [profile]
+        mock_auth_svc.get_current_profile.return_value = profile
+        mock_auth_svc.env = SimpleNamespace(requires_auth=True)
+        mock_service.current_auth_service.return_value = mock_auth_svc
+        result = runner.invoke(app, ["auth", "list", "-o", "json"])
+    assert result.exit_code == 0, result.output
+    data = json.loads(result.output)
+    assert isinstance(data, list)
+    assert data[0]["name"] == "prof"
+    # Secrets must not leak through.
+    assert "api_key" not in data[0]
 
 
 def test_auth_env_list_json(patched_auth: Any) -> None:
